refactor: move constants to utils folder
